requirements.txt
requests
google-generativeai
translators
